クラス名を詳細化
